+++ test/testutils.ts
@@ -64,5 +64,3 @@ export const getSyncPromiseRejectOnFirstCall = <Y extends any[]>(reason: unknown
     }
   });
 };
-
-export const flushPromises = (): Promise<void> => new Promise(setImmediate);